Release 6.3.0
* FslDataBinding now supports two way binding.
* UI.Benchmark now supports a draw cache option.
* FslSimpleUI
  * Draw caching support.
  * Handles SRGB and HDR framebuffers better.
* Added examples for Custom app services.
* Android HDR check improved.
* FslNativeWindowSystem modernized.
* Yocto: Improved Wayland XDG implementations.
  * Now supports XDG decorations where applicable.
* FslBuild
  * Visual Studio projects no longer references unused flavors.
  * Clang format/tidy are now flavor aware.
  * -open2 can now supply arguments to visual studio code (can be used as a workaround for wayland issues).
* Upgraded fmt to 10.1.1
* Upgraded clang tools to 16

+++ DemoFramework/FslBase/UnitTest/source/FslBase/UnitTest/Math/Pixel/Test_PxThicknessU16.cpp
@@ -71,8 +71,8 @@ TEST(TestMathPixel_PxThicknessU16, Values)
   EXPECT_EQ(top, value.Top);
   EXPECT_EQ(right, value.Right);
   EXPECT_EQ(bottom, value.Bottom);
-  EXPECT_EQ((left.Value + right.Value), value.SumX().Value);
-  EXPECT_EQ((top.Value + bottom.Value), value.SumY().Value);
+  EXPECT_EQ(static_cast<uint64_t>(left.Value + right.Value), value.SumX().Value);
+  EXPECT_EQ(static_cast<uint64_t>(top.Value + bottom.Value), value.SumY().Value);
 }
 
 TEST(TestMathPixel_PxThicknessU16, OperatorEqual)
